+++ src/main/java/com/inz/inz/adapter/ReportAdapter.java
@@ -3,8 +3,10 @@ package com.inz.inz.adapter;
 import com.inz.inz.ExcpetionHandler.AuthenticationException;
 import com.inz.inz.ExcpetionHandler.DbException;
 import com.inz.inz.ExcpetionHandler.EnumExcpetion;
+import com.inz.inz.ExcpetionHandler.ExceptionModel;
 import com.inz.inz.entity.ReportEntity;
 import com.inz.inz.resoruce.MarkResourcePost;
+import com.inz.inz.resoruce.NotActiveResource;
 import com.inz.inz.resoruce.ReportResource;
 import com.inz.inz.resoruce.ReportResourcePost;
 
@@ -17,4 +19,8 @@ public interface ReportAdapter {
     ReportResource getReport(Long id) throws DbException;
 
     void addMArk(MarkResourcePost markResourcePost) throws DbException, AuthenticationException;
+
+    void markAsNotActive(NotActiveResource notActiveResource) throws ExceptionModel, AuthenticationException;
+
+    void markAsFalse(NotActiveResource notActiveResource) throws AuthenticationException, DbException;
 }