+++ src/main/java/com/inz/inz/scheduler/BanScheduler.java
@@ -1,0 +1,45 @@
+package com.inz.inz.scheduler;
+
+import com.inz.inz.repository.BanEntityRepository;
+import com.inz.inz.repository.UserRepository;
+import com.inz.inz.seciurity.model.User;
+import org.joda.time.DateTime;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.scheduling.annotation.Scheduled;
+
+import java.util.Date;
+import java.util.List;
+
+@Configuration
+@EnableScheduling
+public class BanScheduler {
+
+    @Autowired
+    UserRepository userRepository;
+    @Autowired
+    BanEntityRepository banEntityRepository;
+
+    //0 0 12 * * ?
+    @Scheduled(cron = "0 * * ? * *")
+    public void checkBans() {
+        List<User> users = userRepository.findAll();
+        if (users != null) {
+            users.stream().forEach(x -> {
+                if (x.getBanEntity().isBanned()) {
+                    Date date = x.getBanEntity().getDate();
+                    DateTime dateTime = new DateTime(date);
+                    if (dateTime.isBeforeNow()) {
+                        x.getBanEntity().setDate(null);
+                        x.getBanEntity().setBanned(false);
+                        banEntityRepository.save(x.getBanEntity());
+                        userRepository.save(x);
+                    }
+                }
+            });
+        }
+
+        System.out.println(users.stream().filter(x -> x.getBanEntity().isBanned()).count());
+    }
+}